requirements.txt
python-docx
webvtt-py
opencv-python
openai
numpy
python-dotenv
